requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
openai
python-dotenv
